Do not rely on atof() for version detection
Removes atof() in the AGI version detection. atof() is locale dependent.
If the user has a locale were the decimal point is something other than
'.' the version detection will fail.

On Linux this can be tested by running:
    LC_NUMERIC="sv_SE.UTF-8" AGIStudio
A 2.917 game will then be detected as a 2.0 game.

+++ src/agicommands.h
@@ -48,6 +48,6 @@ extern CommandStruct TestCommand[19];  //tests for different flags, etc
 extern CommandStruct AGICommand[182];  //all the other
 
 
-extern void  CorrectCommands(double VerNum);
+extern void  CorrectCommands(long VerNum);
 
 #endif
